BirthdayReminderBot 6.0.20221006 + отправление поздравления в чат + редактирование записи (добавление чата/чатов, куда направлять поздравление + проверка новой записи на правильность ввода + getid для проверки id нужного чата

+++ BirthdayReminder.py
@@ -6,7 +6,9 @@ import datetime
 
 print('@BirthdayReminderBot запущен')
 
-bot = telebot.TeleBot("5407469548:AAH66oGKqUE5PWm-dOGCixXk0sZWRMlyglE")  # токен бота 5464014913:AAEW7TYzUvNurSjsIT4xwuwf7KOKogmODIQ
+# bot = telebot.TeleBot("5407469548:AAHpPNs0W8_4DWOUP3gNm1wtIkKnACxp9iY")  # Тестовый БОТ Super Bot
+bot = telebot.TeleBot("5464014913:AAEW7TYzUvNurSjsIT4xwuwf7KOKogmODIQ")  # Рабочий BirthdayReminderBot
+
 
 new_entry_id = ''
 new_birthday_name = ''
@@ -56,6 +58,19 @@ def start(message):
         bot.send_message(message.from_user.id, 'Скопируй из спика и пришли номер id записи, которую нужно удалить.')
         my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + '/del - Start')
         bot.register_next_step_handler(message, del_entry)  # следующий шаг – функция del_entry
+    elif message.text == '/edit':
+        bot.send_message(message.from_user.id, 'Для отправки поздравления в чат скопируй из спика и пришли номер id '
+                                               'записи и id чата. В формате id_записи/id_чата1, id_чата2, id_чата3')
+        bot.send_message(message.from_user.id, 'если вы не знаете id чата воспользуйтесь командой /getid')
+        my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + '/edit - Start')
+        bot.register_next_step_handler(message, edit_entry)  # следующий шаг – функция edit_entry
+    elif message.text == '/getid':
+        bot.send_message(message.from_user.id, 'Перешли сообщение юзера или чата, ID которого хочешь узнать'
+                                               '(Если ты хочешь узнать ID чата, то пересылай именно сообщения чата, '
+                                               'не людей, которые пишут в этом чате')
+        my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + '/getid - Start')
+
+        bot.register_next_step_handler(message, get_chat_id)  # следующий шаг – функция get_chat_id
 
     else:
         bot.send_message(message.from_user.id, help_message)
@@ -63,19 +78,45 @@ def start(message):
     my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'start - Done')
 
 
+# добавление новой записи
 def add_new_entry(message):
     global new_birthday_name, new_birthday_date, new_remind_or_not, new_reminder_period, new_user_name
 
     my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'add_new_entry - Start')
 
     try:
+        if '/' not in str(message.text):
+            bot.send_message(message.from_user.id, 'Не хватает разделителя "/" между именем и датой')
+            raise ZeroDivisionError
+
         new_entry = str(message.text).split('/')
         new_birthday_name = new_entry[0]
         new_birthday_date = new_entry[1]
 
-        # if new_birthday_name == '':
-        # bot.send_message(message.from_user.id, 'Имя не заполнено')
-        # bot.register_next_step_handler(message, start)  # следующий шаг – функция start
+        if new_birthday_name == '':
+            if new_birthday_date.isalpha():
+                raise IndexError
+            else:
+                bot.send_message(message.from_user.id, 'Имя не может быть пустым')
+                raise ZeroDivisionError
+        elif '.' not in new_birthday_date:
+            bot.send_message(message.from_user.id, 'Формат даты должен быть записан через точку, сначала день, '
+                                                   'затем месяц. Например "31.12" - 31е Декабря')
+            raise ZeroDivisionError
+
+        ddmm = str(new_birthday_date).split('.')
+        dd = ddmm[0]
+        mm = ddmm[1]
+
+        if len(dd) != 2 or len(mm) != 2:
+            bot.send_message(message.from_user.id, 'Дата должна иметь вид "ДД.ММ" в днях и месяце по две цифры, '
+                                                   'например: "03.03" (3е марта)')
+            raise ZeroDivisionError
+
+        elif int(dd) > 31 or int(mm) > 12:
+            bot.send_message(message.from_user.id,
+                             'Проверьте дату. Дней должно быть не больше 31, а месяц не больше 12')
+            raise ZeroDivisionError
 
         my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'add_new_entry - Processing')
 
@@ -91,13 +132,19 @@ def add_new_entry(message):
 
         bot.send_message(message.from_user.id, text=question, reply_markup=keyboard)
 
+    except ZeroDivisionError:
+        bot.send_message(message.from_user.id, 'Попробуй еще Имя именника/дата рождения ДД.ММ')
+        bot.register_next_step_handler(message, add_new_entry)  # следующий шаг – функция add_new_entry
+
     except:
         bot.send_message(message.from_user.id, error_message)
-        bot.register_next_step_handler(message, start)  # следующий шаг – функция start
+        bot.send_message(message.from_user.id, help_message)
+        # bot.register_next_step_handler(message, start)  # следующий шаг – функция start
 
     my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'add_new_entry - PreDone')
 
 
+# Список дней рождений юзера
 def user_list(list_user_id):
     connection_list = pymysql.connect(host='31.31.198.35',
                                       user='u1771772_default',
@@ -118,12 +165,20 @@ def user_list(list_user_id):
                 result = cursor_list.fetchall()
 
                 for line in result:
+
+                    chat_id_text = ''
+
+                    if str(line['chat_id']) != '':
+                        # chat_ids = str(line['chat_id']).split(', ')
+
+                        chat_id_text = '\n💬: ' + str(line['chat_id'])
+
                     list_line = '🆔: `' + str(line['id']) + '`\n' + '👤: *' + str(
-                        line['birthday_name']) + '*\n📆: ' + str(line[
-                                                                     'birthday_date'])
+                        line['birthday_name']) + '*\n📆: ' + str(line['birthday_date']) + str(chat_id_text)
 
                     bot.send_message(user_id, list_line, parse_mode='MarkDown')
-        bot.send_message(user_id, 'Для удаления записи введите /del')
+        bot.send_message(user_id, 'Для удаления записи введите /del\n'
+                                  'Для изменения записи введите /edit')
 
     except:
         bot.send_message(user_id, error_message)
@@ -131,6 +186,7 @@ def user_list(list_user_id):
     my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'user_list - Done')
 
 
+# обработка напоминания о сегодняшнем или предстоящем ДР
 def remind_congratulate():
     my_log('remind_congratulate - Start')
     connection_remind_congratulate = pymysql.connect(host='31.31.198.35',
@@ -140,7 +196,6 @@ def remind_congratulate():
                                                      cursorclass=pymysql.cursors.DictCursor)
 
     now_date = time.strftime('%d.%m', time.localtime())
-
 
     with connection_remind_congratulate:
         with connection_remind_congratulate.cursor() as cursor_remind_congratulate:
@@ -152,13 +207,18 @@ def remind_congratulate():
             for line in result:
                 my_log('remind_congratulate - ' + now_date + ' - Done')
 
-                congratulate_text = 'Сегодня ' + line['birthday_date'] + ' отмечает свой День Рождения *' + line[
+                today_date = 'Сегодня ' + line['birthday_date'] + ' отмечает свой День Рождения *' + line[
                     'birthday_name'] + '*'
 
-                bot.send_message(line['user_id'], congratulate_text, parse_mode='MarkDown')
+                congratulate_text = '*' + line['birthday_name'] + '*, с днём Рождения!'
+
+                bot.send_message(line['user_id'], today_date, parse_mode='MarkDown')
 
                 if line['chat_id'] != '0' or line['chat_id'] != '':
-                    bot.send_message(line['chat_id'], congratulate_text, parse_mode='MarkDown')
+                    print('re')
+                    for chat_id in str(line['chat_id']).split(', '):
+                        print(int(chat_id))
+                        bot.send_message(int(chat_id), congratulate_text, parse_mode='MarkDown')
 
         remind_period = [1, 3, 7]
 
@@ -180,10 +240,10 @@ def remind_congratulate():
 
                     bot.send_message(line['user_id'], congratulate_text, parse_mode='MarkDown')
 
-
     time.sleep(100)
 
 
+# удаление записи по id
 def del_entry(message):
     global user_id, new_user_name
 
@@ -197,6 +257,10 @@ def del_entry(message):
     my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'del_entry - connection - Done')
 
     try:
+
+        if '/' in str(message.text) and str(message.text).split('/')[1].isalpha():
+            raise BaseException
+
         my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'del_entry - connection - try')
 
         with connection_del_entry:
@@ -209,8 +273,8 @@ def del_entry(message):
                         new_user_name) + ': ' + 'del_entry - connection - '
                                                 'with connection_del_entry.cursor() as cursor_del_entry' + message.text)
 
-                sql_del = "DELETE FROM `BirthdayReminderBot` WHERE `id`=%s"
-                cursor_del_entry.execute(sql_del, (message.text,))
+                sql_del = "DELETE FROM `BirthdayReminderBot` WHERE `id`=%s AND `user_id`=%s"
+                cursor_del_entry.execute(sql_del, (message.text, user_id))
 
                 my_log(str(user_id) + ': @' + str(
                     new_user_name) + ': ' + 'del_entry - cursor_del_entry - Done ' + str(message.text) + ' : ' + str(
@@ -225,8 +289,61 @@ def del_entry(message):
 
     except:
         bot.send_message(user_id, error_message)
+        bot.send_message(user_id, help_message)
 
-    my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'user_list - Done')
+
+# редактирование записи по id
+def edit_entry(message):
+    global user_id, new_user_name
+
+    user_id = message.from_user.id
+    message_text = str(message.text).split('/')
+
+    connection_edit_entry = pymysql.connect(host='31.31.198.35',
+                                            user='u1771772_default',
+                                            password='56f6hDDRxt96FSvu',
+                                            database='u1771772_default',
+                                            cursorclass=pymysql.cursors.DictCursor)
+    my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'edit_entry - connection - Done')
+
+    try:
+
+        if '/' in str(message.text) and str(message.text).split('/')[1].isalpha():
+            raise BaseException
+
+        my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'edit_entry - connection - try')
+
+        with connection_edit_entry:
+
+            my_log(
+                str(user_id) + ': @' + str(new_user_name) + ': ' + 'edit_entry - connection - connection_edit_entry')
+
+            with connection_edit_entry.cursor() as cursor_edit_entry:
+                my_log(
+                    str(user_id) + ': @' + str(
+                        new_user_name) + ': ' + 'edit_entry - connection - '
+                                                'with connection_edit_entry.cursor() as cursor_edit_entry' + message.text)
+
+                sql_edit = "UPDATE `BirthdayReminderBot` SET `chat_id` = %s WHERE `id` = %s"
+                cursor_edit_entry.execute(sql_edit, (message_text[1], message_text[0]))
+
+            connection_edit_entry.commit()
+
+            my_log(str(user_id) + ': @' + str(new_user_name) + ': ' + 'edit_entry - Done')
+            bot.send_message(message.from_user.id, 'Скорректировал запись')
+
+            user_list(str(user_id))
+
+    except:
+        bot.send_message(user_id, error_message)
+        bot.send_message(user_id, help_message)
+
+
+def get_chat_id(message):
+    get_chat_id_message = 'Ваш id: `' + str(message.from_user.id) + '`\nПереслано от id: `' + str(
+        message.forward_from_chat.id) + '`'
+
+    bot.send_message(message.from_user.id, get_chat_id_message, parse_mode='MarkDown')
 
 
 @bot.callback_query_handler(func=lambda call: True)
